refactor(client): 重构客户端设置相关代码
- 将 GameSettings 类从 gui.settings 包移动到 settings 包
- 更新相关类的导入路径
- 在 GuiToolboxPlusSelect 类中添加 drawDefaultBackground 方法以覆盖默认背景绘制
