+++ src/main/java/moe/takochan/takotech/client/settings/GameSettings.java
@@ -1,4 +1,4 @@
-package moe.takochan.takotech.client.gui.settings;
+package moe.takochan.takotech.client.settings;
 
 import net.minecraft.client.settings.KeyBinding;
